Move MakeVersionCode to Book
A more natural place for it so BookCompressor does not have to know about WiFi publishing
